requirements.txt
evadb
shutil
pandas
python-dotenv